requirements.txt
alembic==1.13.1
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-CKEditor==0.5.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysql-connector-python==8.3.0
mysql-connector-python-rf==2.2.2
pycparser==2.22
PyMySQL==1.1.0
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
